Rename existing Tabs to TabIcon
