+++ packages/teleport/src/components/TabIcon/TabIcon.tsx
